tests/wpt/web-platform-tests/css/css-color/parsing/system-color-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Module Level 4: parsing system color keywords</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#css-system-colors">
<meta name="assert" content="system color keyword support">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("color", "ActiveText", "activetext");
test_valid_value("color", "ButtonFace", "buttonface");
test_valid_value("color", "ButtonText", "buttontext");
test_valid_value("color", "Canvas", "canvas");
test_valid_value("color", "CanvasText", "canvastext");
test_valid_value("color", "Field", "field");
test_valid_value("color", "FieldText", "fieldtext");
test_valid_value("color", "GrayText", "graytext");
test_valid_value("color", "Highlight", "highlight");
test_valid_value("color", "HighlightText", "highlighttext");
test_valid_value("color", "LinkText", "linktext");
test_valid_value("color", "VisitedText", "visitedtext");
</script>
</body>
</html>
